backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
djoser==2.1.0
djangorestframework==3.12.4
django-filter==22.1.0
gunicorn==20.1.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3.post1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyJWT==2.1.0
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.4
urllib3==1.26.15
webcolors==1.11.1
drf-extra-fields==3.5.0
python-dotenv
